requirements.txt
Flask
Flask-Cors
openai
requests
python-dotenv
